fix(tests): split test file to resolve LOC limit violation
Split test_task_status_dashboard.py to fix CI failure:
- Move status simplification tests to dedicated test_status_simplify.py
- Original file: 699 lines → 588 lines (under 600 line limit)
- New file: 141 lines for issue #1782 specific tests
- All 16 tests pass in both files

This resolves the CI Lint & Test failure caused by test file exceeding
600 line limit with exception granted.

Related: #1782
